Adding JSON file, CRAB configuration files (both normal submission and testing using Bristol lcgce01), and updating Global Tag for 2011 7TeV 5_3_X12Oct2013 data re-reco.
Latest CRAB configuration files for Summer 2012 unfolding jobs with 1GeV binning. Included in Issue #62.

Adding centre of mass energy as an option in NTupleTools_cff.py; modified Global Tag clauses to take account of the CMSSW version and the centre of mass energy.

Removing HBHEFilter from the beginning of the PAT Sequence; adding HBHENoiseFilter to ntuples. Issue #48.

Upating, correcting and tidying up selection files to match the recommendations in the reference twikis. Included in Issue #53.